Split endian_write from endian
Only `endian_write` functions require SDL, splitting them out
will allow us to clean up the dependencies a bit.
