+++ Source/dvlnet/frame_queue.cpp
@@ -6,6 +6,7 @@
 #include "dvlnet/packet.h"
 #include "utils/attributes.h"
 #include "utils/endian.hpp"
+#include "utils/endian_write.hpp"
 
 namespace devilution {
 namespace net {
