+++ Source/engine/load_cl2.hpp
@@ -14,6 +14,7 @@
 #include "mpq/mpq_common.hpp"
 #include "utils/cl2_to_clx.hpp"
 #include "utils/endian.hpp"
+#include "utils/endian_write.hpp"
 #include "utils/pointer_value_union.hpp"
 #include "utils/static_vector.hpp"
 #include "utils/status_macros.hpp"
