+++ Source/utils/cel_to_clx.cpp
@@ -13,6 +13,7 @@
 #include "appfat.h"
 #include "utils/clx_encode.hpp"
 #include "utils/endian.hpp"
+#include "utils/endian_write.hpp"
 
 namespace devilution {
