+++ Source/utils/surface_to_clx.cpp
@@ -6,6 +6,7 @@
 
 #include "utils/clx_encode.hpp"
 #include "utils/endian.hpp"
+#include "utils/endian_write.hpp"
 
 #ifdef DEBUG_SURFACE_TO_CLX_SIZE
 #include <iomanip>
